ProjMgtService-Refactoring: Moving dto package to models, and updated their references.
