+++ src/main/java/com/umssonline/proymgt/models/dto/BacklogDto.java
@@ -1,4 +1,4 @@
-package com.umssonline.proymgt.common.dto;
+package com.umssonline.proymgt.models.dto;
 
 public class BacklogDto {
