+++ src/main/java/com/umssonline/proymgt/models/dto/CreateSprintItemDto.java
@@ -1,27 +1,26 @@
-package com.umssonline.proymgt.common.dto;
+package com.umssonline.proymgt.models.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.umssonline.proymgt.models.Priority;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.Id;
 import javax.validation.constraints.NotNull;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Getter
 @Setter
-public class UpdateSprintItemDto {
+public class CreateSprintItemDto {
 
-    @Id
-    @NotNull
-    private Long id;
     @NotNull
     private String name;
     private String description;
     private Priority priority;
     private Integer estimatedTime;
 
+    @JsonIgnore
+    private LocalDate createdOn = LocalDate.now();
     @JsonIgnore
     private LocalDateTime updatedOn = LocalDateTime.now();
 }
